# storage/api-docs/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "HERA 2.0 Mobile API",
    "version": "2.0.0",
    "description": "API untuk integrasi aplikasi mobile HERA (Hydrology Environmental Real-time Analytics) dengan sistem monitoring kualitas air berbasis IoT. Semua endpoint yang dilindungi memerlukan Bearer Token dari Sanctum.\n\n## Cara Penggunaan\n1. Dapatkan token via `POST /api/mobile/login` atau `POST /api/mobile/register`\n2. Klik tombol **Authorize** di atas dan masukkan token\n3. Coba setiap endpoint langsung dari halaman ini",
    "contact": {
      "name": "Tim HERA 2.0 — Politeknik Negeri Lampung",
      "email": "admin@hera.ac.id"
    }
  },
  "servers": [
    {
      "url": "http://103.217.145.187:8000",
      "description": "HERA 2.0 VPS Production"
    },
    {
      "url": "http://127.0.0.1:8000",
      "description": "Lokal Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "sanctum": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Token",
        "description": "Masukkan token dari endpoint /api/mobile/login atau /api/mobile/register"
      }
    },
    "schemas": {
...
      "SensorReading": {
        "type": "object",
        "properties": {
          "location": { "type": "string", "example": "citarum_node_alpha" },
          "tds": { "type": "number", "example": 450.5 },
          "ph": { "type": "number", "example": 7.2 },
          "ec": { "type": "number", "example": 680.3 },
          "suhu_air": { "type": "number", "example": 28.5 },
          "suhu_lingkungan": { "type": "number", "example": 31.0 },
          "kelembapan": { "type": "number", "example": 74.2 },
          "tegangan": { "type": "number", "example": 3.78 },
          "cr_estimated": { "type": "number", "example": 0.0214 },
          "status": { "type": "string", "enum": ["normal", "warning", "danger"], "example": "normal" },
          "timestamp": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/mobile/sensor/latest": {
      "get": {
        "tags": ["Sensor"],
        "summary": "Data sensor terbaru",
        "description": "Mengembalikan satu pembacaan sensor IoT paling terakhir dari InfluxDB. Data mencakup TDS, pH, EC, suhu, kelembapan, tegangan, estimasi Chromium, dan status kualitas air.",
        "operationId": "sensorLatest",
        "security": [{ "sanctum": [] }],
        "responses": {
          "200": { "description": "Data sensor terbaru", "content": { "application/json": { "schema": { "type": "object", "properties": { "status": { "type": "boolean", "example": true }, "message": { "type": "string", "example": "OK" }, "data": { "$ref": "#/components/schemas/SensorReading" } } } } } },
          "404": { "description": "Belum ada data sensor" },
          "401": { "description": "Token tidak valid" }
        }
      }
    },
    "/api/mobile/sensor/history": {
      "get": {
        "tags": ["Sensor"],
        "summary": "Riwayat pembacaan sensor",
        "description": "Mengambil riwayat pembacaan sensor dari InfluxDB. Mendukung filter tanggal dan status kualitas air dengan offset pagination.",
        "operationId": "sensorHistory",
        "security": [{ "sanctum": [] }],
        "parameters": [
          { "name": "from_date", "in": "query", "description": "Tanggal awal (YYYY-MM-DD)", "schema": { "type": "string", "format": "date", "example": "2026-04-01" } },
          { "name": "to_date", "in": "query", "description": "Tanggal akhir (YYYY-MM-DD)", "schema": { "type": "string", "format": "date", "example": "2026-04-16" } },
          { "name": "status", "in": "query", "description": "Filter status kualitas air", "schema": { "type": "string", "enum": ["normal", "warning", "danger"] } },
          { "name": "limit", "in": "query", "description": "Jumlah item per halaman (default: 50, maks: 500)", "schema": { "type": "integer", "example": 50 } },
          { "name": "page", "in": "query", "description": "Nomor halaman", "schema": { "type": "integer", "example": 1 } }
        ],
        "responses": {
          "200": {
            "description": "Riwayat sensor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "OK" },
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/SensorReading" } },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "current_page": { "type": "integer", "example": 1 },
                        "last_page": { "type": "integer", "example": 5 },
                        "per_page": { "type": "integer", "example": 50 },
                        "total": { "type": "integer", "example": 243 }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Token tidak valid" }
        }
      }
    },
    "/api/mobile/sensor/alerts": {
      "get": {
        "tags": ["Sensor"],
        "summary": "Daftar alert sensor",
        "description": "Mengembalikan maksimal 20 pembacaan sensor terkini yang berstatus `warning` atau `danger`. Berguna untuk notifikasi kondisi air berbahaya.",
        "operationId": "sensorAlerts",
        "security": [{ "sanctum": [] }],
        "responses": {
          "200": { "description": "Daftar alert sensor", "content": { "application/json": { "schema": { "type": "object", "properties": { "status": { "type": "boolean", "example": true }, "data": { "type": "array", "items": { "$ref": "#/components/schemas/SensorReading" } } } } } } },
          "401": { "description": "Token tidak valid" }
        }
      }
    },
    "/api/mobile/sensor/daily-stats": {
      "get": {
        "tags": ["Sensor"],
        "summary": "Statistik sensor harian",
        "description": "Mengembalikan ringkasan statistik sensor hari ini: total pembacaan, jumlah warning, jumlah danger, dan rata-rata estimasi kadar Chromium.",
        "operationId": "sensorDailyStats",
        "security": [{ "sanctum": [] }],
        "responses": {
          "200": {
            "description": "Statistik harian",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "date": { "type": "string", "example": "2026-04-16" },
                        "total": { "type": "integer", "example": 48 },
                        "warning": { "type": "integer", "example": 5 },
                        "danger": { "type": "integer", "example": 2 },
                        "avg_cr": { "type": "number", "example": 0.0213, "description": "Rata-rata estimasi Chromium hari ini (mg/L)" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Token tidak valid" }
        }
      }
    },
...
  }
}